fix(server): return value at end of function
In a release build the assert is a no-op, so some value must be returned at the
end of the function. We don't use a default case in the switch statement
because this way we get a warning in case a new value is added to the enum but
we forget to add it to the switch statement.