feat(server): provide way to construct output with metadata
Instead of relying on a separate call to set_metadata the metadata should be
provided by consumers directly with the ctor.

We keep the set_metadata call for updates to the description and in case
consumers need to adjust other fields later on, but add a warning that
providing the metadata through the ctor should be preferred.

To generate descriptions automatically on ctor call the helper function is made
a free function.