refactor(server): reference globals in display
Remove the separate globals struct from public API. Instead consumers should
own the globals in their own infrastructure. We use the globals struct still
for tests.

We introduce a pointer collection to globals in the Display class. These are
transparently set and unset on creation and deletion of a global.