fix(server): manage output configuration lifetimes through front
We need to be extra careful with output configuration lifetimes. In particular
the lifetime of the manager, the front and the resource are independent.

Once a configuration object is used, the resource is uncoupled from the front
and the front is destroyed as soon as the compositor either reports success or
failure. On the other side the manager needs to know about configurations
before that.

The central piece is the configuration front, so here we now keep references to
the other pieces and if required unset them.