ci: build on frameworks master image
We only really need ECM, so in the past we used the base image and installed
(backwards compatible) stable release versions of ECM directly via pacman. But
with the change to KF6 this package is not compatible anymore, so we should
instead rely on a compiled KF6 version of ECM, which the frameworks image
provides.

We can keep this up on master branch until we do our first release. At this
point in time we need to build again on a separate master and stable image in
theory or move back to installing the ECM package, now with the KF6 version.