feat(server): remove ability to add a socket fd
This is currently not used by any consumer. In theory it could be useful for
implementing the new security context protocol. But once we implement this we
may readd it or do it a different way.

But in any case also creating the wl_display when adding a socket is a somewhat
confusing API interface. So if any we should do it differently in the future.

BREAKING CHANGE: Display::add_socket_fd removed